# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Energy and Water",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/historia": {
            "get": {
                "summary": "Obtener todas las historias",
                "operationId": "getAllHistorias",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "summary": "Agregar una nueva historia",
                "operationId": "addHistoria",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo",
                                    "texto",
                                    "activo",
                                    "imagenes"
                                ],
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    },
                                    "imagenes": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "nombre",
                                                "imagen",
                                                "activo"
                                            ],
                                            "properties": {
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "imagen": {
                                                    "type": "string"
                                                },
                                                "activo": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Historia creada con éxito"
                    },
                    "400": {
                        "description": "Validación fallida"
                    },
                    "500": {
                        "description": "Error al agregar la historia"
                    }
                }
            }
        },
        "/v1/historia/{id}": {
            "get": {
                "summary": "Obtener una historia por ID",
                "operationId": "getHistoriaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la historia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Historia no encontrada"
                    },
                    "500": {
                        "description": "Error al recuperar la historia"
                    }
                }
            },
            "put": {
                "summary": "Actualizar o crear una historia",
                "operationId": "updateHistoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la historia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo",
                                    "texto",
                                    "activo",
                                    "imagenes"
                                ],
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    },
                                    "imagenes": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "nombre",
                                                "imagen",
                                                "activo"
                                            ],
                                            "properties": {
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "imagen": {
                                                    "type": "string"
                                                },
                                                "activo": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Historia actualizada con éxito"
                    },
                    "201": {
                        "description": "Historia creada con éxito"
                    },
                    "404": {
                        "description": "Historia no encontrada"
                    },
                    "500": {
                        "description": "Error al actualizar o crear la historia"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una historia por ID",
                "operationId": "deleteHistoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la historia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historia eliminada con éxito"
                    },
                    "404": {
                        "description": "Historia no encontrada"
                    },
                    "500": {
                        "description": "Error al eliminar la historia"
                    }
                }
            },
            "patch": {
                "summary": "Actualizar parcialmente una historia",
                "operationId": "partialUpdateHistoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la historia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    },
                                    "imagenes": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "imagen": {
                                                    "type": "string"
                                                },
                                                "activo": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Historia actualizada con éxito"
                    },
                    "404": {
                        "description": "Historia no encontrada"
                    },
                    "500": {
                        "description": "Error al actualizar la historia"
                    }
                }
            }
        },
...
    }
}